[Chore] Quill Editor 도입으로 이미지 테이블 제거 및 로직 수정
