+++ backend/src/main/java/com/example/bootshelf/boardsvc/board/model/entity/Board.java
@@ -3,7 +3,6 @@ package com.example.bootshelf.boardsvc.board.model.entity;
 import com.example.bootshelf.boardsvc.boardcategory.model.entity.BoardCategory;
 import com.example.bootshelf.boardsvc.boardcomment.model.entity.BoardComment;
 import com.example.bootshelf.boardsvc.boardhistory.model.entity.BoardHistory;
-import com.example.bootshelf.boardsvc.boardimage.model.entity.BoardImage;
 import com.example.bootshelf.boardsvc.boardscrap.model.entity.BoardScrap;
 import com.example.bootshelf.boardsvc.boardtag.model.entity.BoardTag;
 import com.example.bootshelf.boardsvc.boardup.model.entity.BoardUp;
@@ -39,9 +38,6 @@ public class Board {
 
     @OneToMany(mappedBy = "board", fetch = FetchType.LAZY)
     private List<BoardComment> boardCommentList = new ArrayList<>();
-
-    @OneToMany(mappedBy = "board", fetch = FetchType.LAZY)
-    private List<BoardImage> boardImageList = new ArrayList<>();
 
     @OneToMany(mappedBy = "board", fetch = FetchType.LAZY)
     private List<BoardScrap> boardScrapList = new ArrayList<>();
