+++ backend/src/main/java/com/example/bootshelf/boardsvc/board/model/response/GetReadBoardRes.java
@@ -32,8 +32,6 @@ public class GetReadBoardRes {
 
     private String updatedAt;
 
-    private List<String> boardImageList;
-
     private List<GetListCommentBoardRes> boardCommentList;
 
     private List<String> tagList;
