+++ backend/src/main/java/com/example/bootshelf/boardsvc/board/model/response/PostUploadBoardImageRes.java
@@ -1,0 +1,13 @@
+package com.example.bootshelf.boardsvc.board.model.response;
+
+import lombok.*;
+
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+@Builder
+public class PostUploadBoardImageRes {
+    String imageUrl;
+
+}
