+++ backend/src/main/java/com/example/bootshelf/boardsvc/board/service/BoardService.java
@@ -10,9 +10,6 @@ import com.example.bootshelf.boardsvc.board.model.response.PostCreateBoardRes;
 import com.example.bootshelf.boardsvc.board.repository.BoardRepository;
 import com.example.bootshelf.boardsvc.boardcategory.model.entity.BoardCategory;
 import com.example.bootshelf.boardsvc.boardcomment.model.entity.BoardComment;
-import com.example.bootshelf.boardsvc.boardimage.model.entity.BoardImage;
-import com.example.bootshelf.boardsvc.boardimage.repository.BoardImageRepository;
-import com.example.bootshelf.boardsvc.boardimage.service.BoardImageService;
 import com.example.bootshelf.boardsvc.boardscrap.model.response.GetFindBoardScrapRes;
 import com.example.bootshelf.boardsvc.boardscrap.model.response.GetFindBoardScrapResResult;
 import com.example.bootshelf.boardsvc.boardtag.model.entity.BoardTag;
@@ -54,9 +51,7 @@ public class BoardService {
 
     private final BoardRepository boardRepository;
     private final BoardTagService boardTagService;
-    private final BoardImageService boardImageService;
     private final ReviewRepository reviewRepository;
-    private final BoardImageRepository boardImageRepository;
 
 
     @Transactional(readOnly = false)
@@ -83,17 +78,6 @@ public class BoardService {
                 .build();
 
         board = boardRepository.save(board);
-
-        // 게시글 본문의 html에서 img url을 뽑아내기 위해 jsoup 라이브러리 사용
-        Document doc = Jsoup.parse(request.getBoardContent());
-        Elements images = doc.select("img");
-
-        if (!images.isEmpty()) {
-            for (Element img : images) {
-                String imageUrl = img.attr("src");
-                boardImageService.saveImageUrl(board.getIdx(), imageUrl);
-            }
-        }
 
         if (request.getTagList() != null) {
             boardTagService.saveBoardTag(request.getTagList(), board.getIdx());
@@ -152,20 +136,6 @@ public class BoardService {
             }
         }
 
-//        // 이미지 조회
-//        List<GetListImageBoardRes> getListImageBoardResList = new ArrayList<>();
-//
-//        if (!board.getBoardImageList().isEmpty()) {
-//            for (BoardImage boardImage : board.getBoardImageList()) {
-//                GetListImageBoardRes getListImageReviewRes = GetListImageBoardRes.builder()
-//                        .boardImageIdx(boardImage.getIdx())
-//                        .boardImage(boardImage.getBoardImage())
-//                        .build();
-//
-//                getListImageBoardResList.add(getListImageReviewRes);
-//            }
-//        }
-
         GetReadBoardRes getReadBoardRes = GetReadBoardRes.builder()
                 .idx(board.getIdx())
                 .boardCategoryIdx(board.getBoardCategory().getIdx())
@@ -181,7 +151,6 @@ public class BoardService {
                 .scrapCnt(board.getScrapCnt())
                 .createdAt(board.getCreatedAt())
                 .updatedAt(board.getUpdatedAt())
-//                .boardImageList(getListImageBoardResList)
                 .boardCommentList(getListCommentBoardResList)
                 .build();
 
@@ -255,16 +224,6 @@ public class BoardService {
                     .createdAt(board.getCreatedAt())
                     .build();
 
-//            List<BoardImage> boardImageList = board.getBoardImageList();
-//            List<String> fileNames = new ArrayList<>();
-//            if (!boardImageList.isEmpty()) {
-//                for (BoardImage boardImage : boardImageList) {
-//                    String fileName = boardImage.getBoardImage();
-//                    fileNames.add(fileName);
-//                }
-//                getListBoardRes.setBoardImg(fileNames.get(0));
-//            }
-
             getListBoardResList.add(getListBoardRes);
         }
         BaseRes baseRes = BaseRes.builder()
@@ -312,16 +271,6 @@ public class BoardService {
                     .createdAt(board.getCreatedAt())
                     .updatedAt(board.getUpdatedAt())
                     .build();
-
-//            List<BoardImage> boardImageList = board.getBoardImageList();
-//            List<String> fileNames = new ArrayList<>();
-//            if (!boardImageList.isEmpty()) {
-//                for (BoardImage boardImage : boardImageList) {
-//                    String fileName = boardImage.getBoardImage();
-//                    fileNames.add(fileName);
-//                }
-//                getListBoardRes.setBoardImg(fileNames.get(0));
-//            }
 
             getListBoardResList.add(getListBoardRes);
         }
@@ -376,16 +325,6 @@ public class BoardService {
                     .createdAt(board.getCreatedAt())
                     .updatedAt(board.getUpdatedAt())
                     .build();
-
-//            List<BoardImage> boardImageList = board.getBoardImageList();
-//            List<String> fileNames = new ArrayList<>();
-//            if (!boardImageList.isEmpty()) {
-//                for (BoardImage boardImage : boardImageList) {
-//                    String fileName = boardImage.getBoardImage();
-//                    fileNames.add(fileName);
-//                }
-//                getListBoardRes.setBoardImg(fileNames.get(0));
-//            }
 
             Document doc = Jsoup.parse(board.getBoardContent());
             Elements images = doc.select("img");
@@ -450,15 +389,6 @@ public class BoardService {
                     .updatedAt(board.getUpdatedAt())
                     .build();
 
-//            List<BoardImage> boardImageList = board.getBoardImageList();
-//            List<String> fileNames = new ArrayList<>();
-//            if (!boardImageList.isEmpty()) {
-//                for (BoardImage boardImage : boardImageList) {
-//                    String fileName = boardImage.getBoardImage();
-//                    fileNames.add(fileName);
-//                }
-//                getListBoardRes.setBoardImg(fileNames.get(0));
-//            }
             Document doc = Jsoup.parse(board.getBoardContent());
             Elements images = doc.select("img");
 
@@ -522,15 +452,6 @@ public class BoardService {
                     .updatedAt(board.getUpdatedAt())
                     .build();
 
-//            List<BoardImage> boardImageList = board.getBoardImageList();
-//            List<String> fileNames = new ArrayList<>();
-//            if (!boardImageList.isEmpty()) {
-//                for (BoardImage boardImage : boardImageList) {
-//                    String fileName = boardImage.getBoardImage();
-//                    fileNames.add(fileName);
-//                }
-//                getListBoardRes.setBoardImg(fileNames.get(0));
-//            }
             Document doc = Jsoup.parse(board.getBoardContent());
             Elements images = doc.select("img");
 
@@ -636,15 +557,6 @@ public class BoardService {
                     .userProfileImage(board.getUser().getProfileImage())
                     .build();
 
-//            List<BoardImage> boardImageList = board.getBoardImageList();
-//            List<String> fileNames = new ArrayList<>();
-//            if (!boardImageList.isEmpty()) {
-//                for (BoardImage boardImage : boardImageList) {
-//                    String fileName = boardImage.getBoardImage();
-//                    fileNames.add(fileName);
-//                }
-//                getBoardListByQueryRes.setBoardImg(fileNames.get(0));
-//            }
             Document doc = Jsoup.parse(board.getBoardContent());
             Elements images = doc.select("img");
 
@@ -747,18 +659,6 @@ public class BoardService {
 
         boardTagService.updateBoardTag(patchUpdateBoardReq.getTagList(), patchUpdateBoardReq.getBoardIdx());
 
-        // 게시글 본문의 html에서 img url을 뽑아내기 위해 jsoup 라이브러리 사용
-        Document doc = Jsoup.parse(patchUpdateBoardReq.getBoardContent());
-        Elements images = doc.select("img");
-
-        if (!images.isEmpty()) {
-            boardImageRepository.deleteAllByBoard_idx(board.getIdx());
-            for (Element img : images) {
-                String imageUrl = img.attr("src");
-                boardImageService.saveImageUrl(board.getIdx(), imageUrl);
-            }
-        }
-
         board.setBoardTitle(patchUpdateBoardReq.getBoardTitle());
         board.setBoardContent(patchUpdateBoardReq.getBoardContent());
         board.setUpdatedAt(LocalDateTime.now().format(DateTimeFormatter.ofPattern("yyyy/MM/dd HH:mm:ss")));
@@ -817,20 +717,6 @@ public class BoardService {
             }
         }
 
-        // 이미지 조회
-//        List<GetListImageBoardRes> getListImageBoardResList = new ArrayList<>();
-//
-//        if (!board.getBoardImageList().isEmpty()) {
-//            for (BoardImage boardImage : board.getBoardImageList()) {
-//                GetListImageBoardRes getListImageReviewRes = GetListImageBoardRes.builder()
-//                        .boardImageIdx(boardImage.getIdx())
-//                        .boardImage(boardImage.getBoardImage())
-//                        .build();
-//
-//                getListImageBoardResList.add(getListImageReviewRes);
-//            }
-//        }
-
         GetReadBoardRes getReadBoardRes = GetReadBoardRes.builder()
                 .idx(board.getIdx())
                 .boardCategoryName(board.getBoardCategory().getCategoryName())
@@ -838,7 +724,6 @@ public class BoardService {
                 .boardTitle(board.getBoardTitle())
                 .boardContent(board.getBoardContent())
                 .tagList(tags)
-//                .boardImageList(getListImageBoardResList)
                 .build();
 
         BaseRes baseRes = BaseRes.builder()
@@ -884,18 +769,6 @@ public class BoardService {
                     .updatedAt(board.getUpdatedAt())
                     .type("board")
                     .build();
-
-//            List<BoardImage> boardImageList = board.getBoardImageList();
-//            List<String> fileNames = new ArrayList<>();
-//            if (!boardImageList.isEmpty()) {
-//                for (BoardImage boardImage : boardImageList) {
-//                    String fileName = boardImage.getBoardImage();
-//                    fileNames.add(fileName);
-//                }
-//                getListHotBoardRes.setImage(fileNames.get(0));
-//            }
-//
-//            getListResListHotBoard.add(getListHotBoardRes);
 
             Document doc = Jsoup.parse(board.getBoardContent());
             Elements images = doc.select("img");
@@ -960,16 +833,6 @@ public class BoardService {
                     .updatedAt(board.getUpdatedAt())
                     .type("board")
                     .build();
-
-//            List<BoardImage> boardImageList = board.getBoardImageList();
-//            List<String> fileNames = new ArrayList<>();
-//            if (!boardImageList.isEmpty()) {
-//                for (BoardImage boardImage : boardImageList) {
-//                    String fileName = boardImage.getBoardImage();
-//                    fileNames.add(fileName);
-//                }
-//                getListHotBoardRes.setImage(fileNames.get(0));
-//            }
 
             Document doc = Jsoup.parse(board.getBoardContent());
             Elements images = doc.select("img");
