+++ backend/src/main/java/com/example/bootshelf/reviewsvc/review/model/response/GetReadReviewRes.java
@@ -25,7 +25,6 @@ public class GetReadReviewRes {
     private Integer upCnt;
     private Integer commentCnt;
     private String updatedAt;
-    private List<GetListImageReviewRes> reviewImageList;
     private List<GetListCommentReviewRes> reviewCommentList;
 
 }
