+++ backend/src/main/java/com/example/bootshelf/reviewsvc/review/service/ReviewService.java
@@ -10,9 +10,6 @@ import com.example.bootshelf.reviewsvc.review.model.response.*;
 import com.example.bootshelf.reviewsvc.review.repository.ReviewRepository;
 import com.example.bootshelf.reviewsvc.reviewcategory.model.ReviewCategory;
 import com.example.bootshelf.reviewsvc.reviewcomment.model.entity.ReviewComment;
-import com.example.bootshelf.reviewsvc.reviewimage.model.entity.ReviewImage;
-import com.example.bootshelf.reviewsvc.reviewimage.repository.ReviewImageRepository;
-import com.example.bootshelf.reviewsvc.reviewimage.service.ReviewImageService;
 import com.example.bootshelf.user.model.entity.User;
 import lombok.RequiredArgsConstructor;
 import org.jsoup.Jsoup;
@@ -35,10 +32,7 @@ import java.util.Optional;
 @RequiredArgsConstructor
 public class ReviewService {
 
-    private final ReviewImageService reviewImageService;
     private final ReviewRepository reviewRepository;
-    private final ReviewImageRepository reviewImageRepository;
-
 
     @Transactional(readOnly = false)
     public BaseRes createReview(User user, PostCreateReviewReq postCreateReviewReq) {
@@ -68,16 +62,6 @@ public class ReviewService {
 
         review = reviewRepository.save(review);
 
-        // 게시글 본문의 html에서 img url을 뽑아내기 위해 jsoup 라이브러리 사용
-        Document doc = Jsoup.parse(postCreateReviewReq.getReviewContent());
-        Elements images = doc.select("img");
-
-        if (!images.isEmpty()) {
-            for (Element img : images) {
-                String imageUrl = img.attr("src");
-                reviewImageService.saveImageUrl(review.getIdx(), imageUrl);
-            }
-        }
 
         PostCreateReviewRes postCreateReviewRes = PostCreateReviewRes.builder()
                 .reviewIdx(review.getIdx())
@@ -131,12 +115,6 @@ public class ReviewService {
                     .updatedAt(review.getUpdatedAt())
                     .build();
 
-//            List<ReviewImage> reviewImageList = review.getReviewImageList();
-//            if (!reviewImageList.isEmpty()) {
-//                ReviewImage reviewImage = reviewImageList.get(0);
-//                String image = reviewImage.getReviewImage();
-//                getMyListReviewRes.setReviewImage(image);
-//            }
             getMyListReviewResList.add(getMyListReviewRes);
         }
 
@@ -186,12 +164,6 @@ public class ReviewService {
                     .updatedAt(review.getUpdatedAt())
                     .build();
 
-//            List<ReviewImage> reviewImageList = review.getReviewImageList();
-//            if (!reviewImageList.isEmpty()) {
-//                ReviewImage reviewImage = reviewImageList.get(0);
-//                String image = reviewImage.getReviewImage();
-//                getListReviewRes.setReviewImage(image);
-//            }
             Document doc = Jsoup.parse(review.getReviewContent());
             Elements images = doc.select("img");
 
@@ -245,20 +217,6 @@ public class ReviewService {
             }
         }
 
-//        // 이미지 조회
-//        List<GetListImageReviewRes> getListImageReviewResList = new ArrayList<>();
-//
-//        if (!review.getReviewImageList().isEmpty()) {
-//            for (ReviewImage reviewImage : review.getReviewImageList()) {
-//                GetListImageReviewRes getListImageReviewRes = GetListImageReviewRes.builder()
-//                        .reviewImageIdx(reviewImage.getIdx())
-//                        .reviewImage(reviewImage.getReviewImage())
-//                        .build();
-//
-//                getListImageReviewResList.add(getListImageReviewRes);
-//            }
-//        }
-
         GetReadReviewRes getReadReviewRes = GetReadReviewRes.builder()
                 .reviewIdx(review.getIdx())
                 .reviewCategoryName(review.getReviewCategory().getCategoryName())
@@ -273,7 +231,6 @@ public class ReviewService {
                 .scrapCnt(review.getScrapCnt())
                 .commentCnt(review.getCommentCnt())
                 .updatedAt(review.getUpdatedAt())
-//                .reviewImageList(getListImageReviewResList)
                 .reviewCommentList(getListCommentReviewResList)
                 .build();
 
@@ -372,19 +329,6 @@ public class ReviewService {
 
         review.update(patchUpdateReviewReq);
 
-        // 게시글 본문의 html에서 img url을 뽑아내기 위해 jsoup 라이브러리 사용
-        Document doc = Jsoup.parse(patchUpdateReviewReq.getReviewContent());
-        Elements images = doc.select("img");
-
-        if (!images.isEmpty()) {
-            reviewImageRepository.deleteAllByReview_idx(review.getIdx());
-
-            for (Element img : images) {
-                String imageUrl = img.attr("src");
-                reviewImageService.saveImageUrl(review.getIdx(), imageUrl);
-            }
-        }
-
         review.setUpdatedAt(LocalDateTime.now().format(DateTimeFormatter.ofPattern("yyyy/MM/dd HH:mm:ss")));
         reviewRepository.save(review);
 
@@ -452,12 +396,6 @@ public class ReviewService {
                     .updatedAt(review.getUpdatedAt())
                     .build();
 
-//            List<ReviewImage> reviewImageList = review.getReviewImageList();
-//            if (!reviewImageList.isEmpty()) {
-//                ReviewImage reviewImage = reviewImageList.get(0);
-//                String image = reviewImage.getReviewImage();
-//                getSearchListReviewRes.setReviewImage(image);
-//            }
             Document doc = Jsoup.parse(review.getReviewContent());
             Elements images = doc.select("img");
 
@@ -516,12 +454,7 @@ public class ReviewService {
                     .type("review")
                     .build();
 
-//            List<ReviewImage> reviewImageList = review.getReviewImageList();
-//            if (!reviewImageList.isEmpty()) {
-//                ReviewImage reviewImage = reviewImageList.get(0);
-//                String image = reviewImage.getReviewImage();
-//                getListHotReviewRes.setImage(image);
-//            }
+
             Document doc = Jsoup.parse(review.getReviewContent());
             Elements images = doc.select("img");
 
@@ -579,12 +512,6 @@ public class ReviewService {
                     .type("review")
                     .build();
 
-//            List<ReviewImage> reviewImageList = review.getReviewImageList();
-//            if (!reviewImageList.isEmpty()) {
-//                ReviewImage reviewImage = reviewImageList.get(0);
-//                String image = reviewImage.getReviewImage();
-//                getListHotReviewRes.setImage(image);
-//            }
             Document doc = Jsoup.parse(review.getReviewContent());
             Elements images = doc.select("img");
 
@@ -623,27 +550,12 @@ public class ReviewService {
 
         Review review = result.get();
 
-//        List<GetListImageReviewRes> reviewImages = new ArrayList<>();
-//
-//        if (!review.getReviewImageList().isEmpty()) {
-//            for (ReviewImage reviewImage : review.getReviewImageList()) {
-//
-//                GetListImageReviewRes getListImageReviewRes = GetListImageReviewRes.builder()
-//                        .reviewImageIdx(reviewImage.getIdx())
-//                        .reviewImage(reviewImage.getReviewImage())
-//                        .build();
-//
-//                reviewImages.add(getListImageReviewRes);
-//            }
-//        }
-
         GetReadReviewRes getReadReviewRes = GetReadReviewRes.builder()
                 .reviewIdx(review.getIdx())
                 .reviewCategoryName(review.getReviewCategory().getCategoryName())
                 .reviewCategoryIdx(review.getReviewCategory().getIdx())
                 .reviewTitle(review.getReviewTitle())
                 .reviewContent(review.getReviewContent())
-//                .reviewImageList(reviewImages)
                 .courseName(review.getCourseName())
                 .courseEvaluation(review.getCourseEvaluation())
                 .build();
