--- backend/src/main/java/com/example/bootshelf/reviewsvc/reviewimage/model/response/PostUploadReviewImageRes.java
@@ -1,13 +1,0 @@
-package com.example.bootshelf.reviewsvc.reviewimage.model.response;
-
-import lombok.*;
-
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-public class PostUploadReviewImageRes {
-
-    String imageUrl;
-}
